Move facade method to base class
